+++ src/lib.rs
@@ -9,3 +9,4 @@ mod packet;
 mod client;
 mod client_identifier;
 mod error;
+mod keep_alive;
